Isolate IterationType tests
